Added test cases for CSE

+++ resources/S3-CSE/src/main/java/demo/UploadObjectKMSKey.java
@@ -8,11 +8,11 @@ import java.security.spec.X509EncodedKeySpec;
 import com.amazonaws.AmazonServiceException;
 import com.amazonaws.auth.profile.ProfileCredentialsProvider;
 import com.amazonaws.regions.RegionUtils;
-import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3Encryption;
-import com.amazonaws.services.s3.AmazonS3EncryptionClient;
 import com.amazonaws.services.s3.AmazonS3EncryptionClientBuilder;
 import com.amazonaws.services.s3.model.*;
+
+import org.apache.commons.cli.*;
 
 import javax.crypto.KeyGenerator;
 import javax.crypto.SecretKey;
@@ -20,12 +20,10 @@ import javax.crypto.spec.SecretKeySpec;
 
 
 public class UploadObjectKMSKey {
-    private static void kms_crypto(String clientRegion, String bucketName, String keyName) {
-        String kmsKeyId = System.getenv("KMS_ID");
-        boolean authenticated_crypto = System.getenv("AUTHENTICATED_CRYPTO") != null;
-
+    private static void kmsCrypto(String clientRegion, String bucketName, String keyName, String kmsKeyId,
+                                  boolean authenticatedCrypto, boolean getOnly, boolean putOnly) {
         System.out.println("KMS Key: " + kmsKeyId);
-        System.out.println("Authenticated Encryption: " + authenticated_crypto);
+        System.out.println("Authenticated Encryption: " + authenticatedCrypto);
 
         try {
             // Create the encryption client.
@@ -33,7 +31,7 @@ public class UploadObjectKMSKey {
 
             CryptoConfiguration cryptoConfig;
 
-            if (authenticated_crypto) {
+            if (authenticatedCrypto) {
                 // This does AES/GCM/NoPadding
                 cryptoConfig = new CryptoConfiguration(CryptoMode.AuthenticatedEncryption).withAwsKmsRegion(RegionUtils.getRegion(clientRegion));
             } else {
@@ -50,34 +48,33 @@ public class UploadObjectKMSKey {
             // Upload an object using the encryption client.
             String origContent = "S3 Encrypted Object Using KMS-Managed Customer Master Key.";
             int origContentLength = origContent.length();
-            // encryptionClient.putObject(bucketName, keyName, origContent);
 
-            // Download the object. The downloaded object is still encrypted.
-            //S3Object downloadedObject = encryptionClient.getObject(bucketName, keyName);
-
-            GetObjectRequest a = new GetObjectRequest(bucketName, keyName);
-            a.setRange(3);
-
-            S3Object downloadedObject = encryptionClient.getObject(a);
-
-
-            S3ObjectInputStream input = downloadedObject.getObjectContent();
-
-            // Decrypt and read the object and close the input stream.
-            byte[] readBuffer = new byte[4096];
-            ByteArrayOutputStream baos = new ByteArrayOutputStream(4096);
-            int bytesRead = 0;
-            int decryptedContentLength = 0;
-
-            while ((bytesRead = input.read(readBuffer)) != -1) {
-                baos.write(readBuffer, 0, bytesRead);
-                decryptedContentLength += bytesRead;
+            if (putOnly || !getOnly) {
+                encryptionClient.putObject(bucketName, keyName, origContent);
             }
-            input.close();
 
-            // Verify that the original and decrypted contents are the same size.
+            if (getOnly || !putOnly) {
+                // Download the object. The downloaded object is still encrypted.
+                S3Object downloadedObject = encryptionClient.getObject(bucketName, keyName);
+                S3ObjectInputStream input = downloadedObject.getObjectContent();
+
+                // Decrypt and read the object and close the input stream.
+                byte[] readBuffer = new byte[4096];
+                ByteArrayOutputStream baos = new ByteArrayOutputStream(4096);
+                int bytesRead = 0;
+                int decryptedContentLength = 0;
+
+                while ((bytesRead = input.read(readBuffer)) != -1) {
+                    baos.write(readBuffer, 0, bytesRead);
+                    decryptedContentLength += bytesRead;
+                }
+                input.close();
+
+                // Verify that the original and decrypted contents are the same size.
+
+                System.out.println("Decrypted content length: " + decryptedContentLength);
+            }
             System.out.println("Original content length: " + origContentLength);
-            System.out.println("Decrypted content length: " + decryptedContentLength);
         }
         catch(AmazonServiceException e) {
             // The call was transmitted successfully, but Amazon S3 couldn't process
@@ -92,18 +89,19 @@ public class UploadObjectKMSKey {
 
     }
 
-    private static void symmetric_crypto(String clientRegion, String bucketName, String keyName) throws Exception {
-        String masterKeyDir = System.getenv("KEY_DIR");
+    private static void symmetricCrypto(String clientRegion, String bucketName, String keyName,
+                                        String masterKeyDir, boolean getOnly, boolean putOnly) throws Exception {
         String masterKeyName = "secret.key";
 
         System.out.println("symmetric Encryption");
 
         KeyGenerator symKeyGenerator = KeyGenerator.getInstance("AES");
         symKeyGenerator.init(256);
-        // SecretKey symKey = symKeyGenerator.generateKey();
+        SecretKey symKey = symKeyGenerator.generateKey();
 
-        // saveSymmetricKey(masterKeyDir, masterKeyName, symKey);
-        SecretKey symKey = loadSymmetricAESKey(masterKeyDir, masterKeyName, "AES");
+        // Only saves if key doesnt already exist
+        saveSymmetricKey(masterKeyDir, masterKeyName, symKey);
+        symKey = loadSymmetricAESKey(masterKeyDir, masterKeyName, "AES");
 
         try {
             EncryptionMaterials encryptionMaterials = new EncryptionMaterials(symKey);
@@ -116,27 +114,32 @@ public class UploadObjectKMSKey {
             // Upload an object using the encryption client.
             String origContent = "S3 Encrypted Object Using symmetric.";
             int origContentLength = origContent.length();
-            encryptionClient.putObject(bucketName, keyName, origContent);
+            if (putOnly || !getOnly) {
+                encryptionClient.putObject(bucketName, keyName, origContent);
+            }
 
             // Download the object. The downloaded object is still encrypted.
-            S3Object downloadedObject = encryptionClient.getObject(bucketName, keyName);
-            S3ObjectInputStream input = downloadedObject.getObjectContent();
+            if (getOnly || !putOnly) {
+                S3Object downloadedObject = encryptionClient.getObject(bucketName, keyName);
+                S3ObjectInputStream input = downloadedObject.getObjectContent();
 
-            // Decrypt and read the object and close the input stream.
-            byte[] readBuffer = new byte[4096];
-            ByteArrayOutputStream baos = new ByteArrayOutputStream(4096);
-            int bytesRead = 0;
-            int decryptedContentLength = 0;
+                // Decrypt and read the object and close the input stream.
+                byte[] readBuffer = new byte[4096];
+                ByteArrayOutputStream baos = new ByteArrayOutputStream(4096);
+                int bytesRead = 0;
+                int decryptedContentLength = 0;
 
-            while ((bytesRead = input.read(readBuffer)) != -1) {
-                baos.write(readBuffer, 0, bytesRead);
-                decryptedContentLength += bytesRead;
+                while ((bytesRead = input.read(readBuffer)) != -1) {
+                    baos.write(readBuffer, 0, bytesRead);
+                    decryptedContentLength += bytesRead;
+                }
+                input.close();
+
+                // Verify that the original and decrypted contents are the same size.
+
+                System.out.println("Decrypted content length: " + decryptedContentLength);
             }
-            input.close();
-
-            // Verify that the original and decrypted contents are the same size.
             System.out.println("Original content length: " + origContentLength);
-            System.out.println("Decrypted content length: " + decryptedContentLength);
         }
         catch(AmazonServiceException e) {
             // The call was transmitted successfully, but Amazon S3 couldn't process
@@ -151,8 +154,8 @@ public class UploadObjectKMSKey {
 
     }
 
-    private static void asymmetric_crypto(String clientRegion, String bucketName, String keyName) throws Exception {
-        String masterKeyDir = System.getenv("KEY_DIR");
+    private static void asymmetricCrypto(String clientRegion, String bucketName, String keyName,
+                                         String masterKeyDir, boolean getOnly, boolean putOnly) throws Exception {
         String pubKeyName = "secret.pub";
         String privKeyName = "secret.priv";
 
@@ -177,29 +180,32 @@ public class UploadObjectKMSKey {
             // Upload an object using the encryption client.
             String origContent = "S3 Encrypted Object Using asymmetric encryption.";
             int origContentLength = origContent.length();
-            encryptionClient.putObject(bucketName, keyName, origContent);
-
+            if (putOnly || !getOnly) {
+                encryptionClient.putObject(bucketName, keyName, origContent);
+            }
             // Download the object. The downloaded object is still encrypted.
 
+            if (getOnly || !putOnly) {
+                S3Object downloadedObject = encryptionClient.getObject(bucketName, keyName);
+                S3ObjectInputStream input = downloadedObject.getObjectContent();
 
-            S3Object downloadedObject = encryptionClient.getObject(bucketName, keyName);
-            S3ObjectInputStream input = downloadedObject.getObjectContent();
+                // Decrypt and read the object and close the input stream.
+                byte[] readBuffer = new byte[4096];
+                ByteArrayOutputStream baos = new ByteArrayOutputStream(4096);
+                int bytesRead = 0;
+                int decryptedContentLength = 0;
 
-            // Decrypt and read the object and close the input stream.
-            byte[] readBuffer = new byte[4096];
-            ByteArrayOutputStream baos = new ByteArrayOutputStream(4096);
-            int bytesRead = 0;
-            int decryptedContentLength = 0;
+                while ((bytesRead = input.read(readBuffer)) != -1) {
+                    baos.write(readBuffer, 0, bytesRead);
+                    decryptedContentLength += bytesRead;
+                }
+                input.close();
 
-            while ((bytesRead = input.read(readBuffer)) != -1) {
-                baos.write(readBuffer, 0, bytesRead);
-                decryptedContentLength += bytesRead;
+                // Verify that the original and decrypted contents are the same size.
+
+                System.out.println("Decrypted content length: " + decryptedContentLength);
             }
-            input.close();
-
-            // Verify that the original and decrypted contents are the same size.
             System.out.println("Original content length: " + origContentLength);
-            System.out.println("Decrypted content length: " + decryptedContentLength);
         }
         catch(AmazonServiceException e) {
             // The call was transmitted successfully, but Amazon S3 couldn't process
@@ -215,10 +221,14 @@ public class UploadObjectKMSKey {
     }
 
     private static void saveSymmetricKey(String masterKeyDir, String masterKeyName, SecretKey secretKey) throws IOException {
-        X509EncodedKeySpec x509EncodedKeySpec = new X509EncodedKeySpec(secretKey.getEncoded());
-        FileOutputStream keyOutputStream = new FileOutputStream(masterKeyDir + File.separator + masterKeyName);
-        keyOutputStream.write(x509EncodedKeySpec.getEncoded());
-        keyOutputStream.close();
+        File outputFile = new File(masterKeyDir + File.separator + masterKeyName);
+
+        if (!outputFile.exists()) {
+            X509EncodedKeySpec x509EncodedKeySpec = new X509EncodedKeySpec(secretKey.getEncoded());
+            FileOutputStream keyOutputStream = new FileOutputStream(masterKeyDir + File.separator + masterKeyName);
+            keyOutputStream.write(x509EncodedKeySpec.getEncoded());
+            keyOutputStream.close();
+        }
     }
 
     private static SecretKey loadSymmetricAESKey(String masterKeyDir, String masterKeyName, String algorithm)
@@ -239,20 +249,24 @@ public class UploadObjectKMSKey {
                                     String publicKeyName,
                                     String privateKeyName,
                                     KeyPair keyPair) throws IOException {
-        PrivateKey privateKey = keyPair.getPrivate();
-        PublicKey publicKey = keyPair.getPublic();
+        File outputFile = new File(dir + File.separator + publicKeyName);
 
-        // Write the public key to the specified file.
-        X509EncodedKeySpec x509EncodedKeySpec = new X509EncodedKeySpec(publicKey.getEncoded());
-        FileOutputStream publicKeyOutputStream = new FileOutputStream(dir + File.separator + publicKeyName);
-        publicKeyOutputStream.write(x509EncodedKeySpec.getEncoded());
-        publicKeyOutputStream.close();
+        if (!outputFile.exists()) {
+            PrivateKey privateKey = keyPair.getPrivate();
+            PublicKey publicKey = keyPair.getPublic();
 
-        // Write the private key to the specified file.
-        PKCS8EncodedKeySpec pkcs8EncodedKeySpec = new PKCS8EncodedKeySpec(privateKey.getEncoded());
-        FileOutputStream privateKeyOutputStream = new FileOutputStream(dir + File.separator + privateKeyName);
-        privateKeyOutputStream.write(pkcs8EncodedKeySpec.getEncoded());
-        privateKeyOutputStream.close();
+            // Write the public key to the specified file.
+            X509EncodedKeySpec x509EncodedKeySpec = new X509EncodedKeySpec(publicKey.getEncoded());
+            FileOutputStream publicKeyOutputStream = new FileOutputStream(dir + File.separator + publicKeyName);
+            publicKeyOutputStream.write(x509EncodedKeySpec.getEncoded());
+            publicKeyOutputStream.close();
+
+            // Write the private key to the specified file.
+            PKCS8EncodedKeySpec pkcs8EncodedKeySpec = new PKCS8EncodedKeySpec(privateKey.getEncoded());
+            FileOutputStream privateKeyOutputStream = new FileOutputStream(dir + File.separator + privateKeyName);
+            privateKeyOutputStream.write(pkcs8EncodedKeySpec.getEncoded());
+            privateKeyOutputStream.close();
+        }
     }
 
     private static KeyPair loadKeyPair(String dir,
@@ -287,10 +301,57 @@ public class UploadObjectKMSKey {
 
 
     public static void main(String[] args) throws IOException {
-        String bucketName = System.getenv("S3_BUCKET_NAME");
-        String keyName = System.getenv("S3_KEY");
-        String clientRegion = System.getenv("REGION");
-        String cryptoType = System.getenv("CRYPTO_TYPE");
+        Options options = new Options();
+        Option bucketNameInput = new Option("b", "bucket-name", true, "S3 Bucket to use");
+        bucketNameInput.setRequired(true);
+        options.addOption(bucketNameInput);
+
+        Option keyNameInput = new Option("k", "key-name", true, "S3 File name to use");
+        keyNameInput.setRequired(true);
+        options.addOption(keyNameInput);
+
+        Option regionInput = new Option("r", "region", true, "AWS Region");
+        regionInput.setRequired(true);
+        options.addOption(regionInput);
+
+        Option cryptoTypeInput = new Option("c", "crypto-type", true, "Which crypto type to use");
+        cryptoTypeInput.setRequired(true);
+        options.addOption(cryptoTypeInput);
+
+        Option kmsKeyIdInput = new Option("a", "kms-key-id", true, "KMS Key ID or alias");
+        options.addOption(kmsKeyIdInput);
+
+        Option authenticatedCryptoInput = new Option("d", "authenticated-crypto", false, "Authenticated crypto");
+        options.addOption(authenticatedCryptoInput);
+
+        Option keyDirInput = new Option("e", "key-dir", true, "Symmetric or Asymmetric key directory");
+        options.addOption(keyDirInput);
+
+        Option putOnlyInput = new Option("p", "put-only", false, "Upload only");
+        options.addOption(putOnlyInput);
+
+        Option getOnlyInput = new Option("g", "get-only", false, "Download only");
+        options.addOption(getOnlyInput);
+
+        CommandLineParser parser = new DefaultParser();
+        HelpFormatter formatter = new HelpFormatter();
+        CommandLine cmd = null;
+
+        try {
+            cmd = parser.parse(options, args);
+        } catch (ParseException e) {
+            System.out.println(e.getMessage());
+            formatter.printHelp("s3_cse_example", options);
+
+            System.exit(1);
+        }
+
+        String bucketName = cmd.getOptionValue("bucket-name");
+        String keyName = cmd.getOptionValue("key-name");
+        String clientRegion = cmd.getOptionValue("region");
+        String cryptoType = cmd.getOptionValue("crypto-type");
+        boolean getOnly = cmd.hasOption("get-only");
+        boolean putOnly = cmd.hasOption("put-only");
 
         System.out.println("S3: s3://" + bucketName + "/" + keyName);
         System.out.println("Region: " + clientRegion);
@@ -298,13 +359,20 @@ public class UploadObjectKMSKey {
         try {
             switch (cryptoType) {
                 case "kms":
-                    UploadObjectKMSKey.kms_crypto(clientRegion, bucketName, keyName);
+                    String kmsKeyId = cmd.getOptionValue("kms-key-id");
+                    boolean authenticatedCrypto = cmd.hasOption("authenticated-crypto");
+
+                    UploadObjectKMSKey.kmsCrypto(clientRegion, bucketName, keyName, kmsKeyId, authenticatedCrypto, getOnly, putOnly);
                     break;
                 case "asymmetric":
-                    UploadObjectKMSKey.asymmetric_crypto(clientRegion, bucketName, keyName);
+                    String keyDir = cmd.getOptionValue("key-dir");
+
+                    UploadObjectKMSKey.asymmetricCrypto(clientRegion, bucketName, keyName, keyDir, getOnly, putOnly);
                     break;
                 case "symmetric":
-                    UploadObjectKMSKey.symmetric_crypto(clientRegion, bucketName, keyName);
+                    keyDir = cmd.getOptionValue("key-dir");
+
+                    UploadObjectKMSKey.symmetricCrypto(clientRegion, bucketName, keyName, keyDir, getOnly, putOnly);
                     break;
 
                 default:
